moved handleNextClick, handleCheckClick to CheckButton.js
